@@ -3,7 +3,7 @@ publish_to: none
 
 # Tests require flutter beta or greater to run.
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <3.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
