@@ -7,8 +7,7 @@
 import 'package:flutter/material.dart';
 
 abstract class GoogleMapExampleAppPage extends StatelessWidget {
-  const GoogleMapExampleAppPage(this.leading, this.title, {Key? key})
-      : super(key: key);
+  const GoogleMapExampleAppPage(this.leading, this.title, {super.key});
 
   final Widget leading;
   final String title;
