@@ -22,7 +22,7 @@ class MarkerIconsPage extends GoogleMapExampleAppPage {
 }
 
 class MarkerIconsBody extends StatefulWidget {
-  const MarkerIconsBody({Key? key}) : super(key: key);
+  const MarkerIconsBody({super.key});
 
   @override
   State<StatefulWidget> createState() => MarkerIconsBodyState();
