@@ -23,7 +23,7 @@ void main() {
 // #enddocregion DisplayMode
 
 class MyApp extends StatefulWidget {
-  const MyApp({Key? key}) : super(key: key);
+  const MyApp({super.key});
 
   @override
   State<MyApp> createState() => _MyAppState();
