@@ -109,7 +109,7 @@ class InfoWindow {
 @immutable
 class MarkerId extends MapsObjectId<Marker> {
   /// Creates an immutable identifier for a [Marker].
-  const MarkerId(String value) : super(value);
+  const MarkerId(super.value);
 }
 
 /// Marks a geographical location on the map.
