@@ -10,8 +10,8 @@ import 'types.dart';
 // (Do not re-export)
 class PolygonUpdates extends MapsObjectUpdates<Polygon> {
   /// Computes [PolygonUpdates] given previous and current [Polygon]s.
-  PolygonUpdates.from(Set<Polygon> previous, Set<Polygon> current)
-      : super.from(previous, current, objectName: 'polygon');
+  PolygonUpdates.from(super.previous, super.current)
+      : super.from(objectName: 'polygon');
 
   /// Set of Polygons to be added in this update.
   Set<Polygon> get polygonsToAdd => objectsToAdd;
