@@ -10,9 +10,8 @@ import 'package:google_maps_flutter_platform_interface/src/types/utils/maps_obje
 import 'test_maps_object.dart';
 
 class TestMapsObjectUpdate extends MapsObjectUpdates<TestMapsObject> {
-  TestMapsObjectUpdate.from(
-      Set<TestMapsObject> previous, Set<TestMapsObject> current)
-      : super.from(previous, current, objectName: 'testObject');
+  TestMapsObjectUpdate.from(super.previous, super.current)
+      : super.from(objectName: 'testObject');
 }
 
 void main() {
